requirements.txt
keboola.component==1.4.4
keboola.utils==1.1.0
openai==1.67.0
pinecone==6.0.2
python-dotenv==1.0.1
scipy==1.15.2
streamlit==1.39.0
uuid
